chore(loader): splits loader to two files
